moved integration to correct folders "mostly client only"
